+++ Sources/SwineCraft_XP/Types/UnsignedVarLong.swift
@@ -28,7 +28,7 @@ struct UnsignedVarLong: ExpressibleByIntegerLiteral, MCPEPacketEncodable {
 
     func encode(_ buf: inout NIOCore.ByteBuffer) throws {
         guard self.backingInt != 0 else {
-            buf.writeInteger(self.backingInt)
+            buf.writeInteger(UInt8(self.backingInt))
 
             return
         }
